move SVG code to its own directory
